# backend/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"database/sql"
	"time"
)

type Challenge struct {
	ID          int32        `json:"id"`
	CtfID       int32        `json:"ctf_id"`
	Name        string       `json:"name"`
	Description string       `json:"description"`
	Flag        string       `json:"flag"`
	CreatedAt   sql.NullTime `json:"created_at"`
	HedgedocUrl string       `json:"hedgedoc_url"`
}

type Ctf struct {
	ID          int32     `json:"id"`
	Phrase      string    `json:"phrase"`
	Name        string    `json:"name"`
	Description string    `json:"description"`
	StartDate   time.Time `json:"start_date"`
	EndDate     time.Time `json:"end_date"`
	AuthorID    int32     `json:"author_id"`
}

type Guest struct {
	ID        int32        `json:"id"`
	Nickname  string       `json:"nickname"`
	LastSeen  sql.NullTime `json:"last_seen"`
	CreatedAt sql.NullTime `json:"created_at"`
	CtfID     int32        `json:"ctf_id"`
}

type User struct {
	ID           int32  `json:"id"`
	Username     string `json:"username"`
	PasswordHash string `json:"password_hash"`
	IsAdmin      bool   `json:"is_admin"`
}

type UserCtf struct {
	UserID   int32        `json:"user_id"`
	CtfID    int32        `json:"ctf_id"`
	JoinedAt sql.NullTime `json:"joined_at"`
}