# backend/docs/swagger.yaml
definitions:
  db.ChangePasswordParams:
    properties:
      password_hash:
        type: string
      username:
        type: string
    type: object
  db.CreateUserParams:
    properties:
      password_hash:
        type: string
      username:
        type: string
    type: object
...
  db.User:
    properties:
      id:
        type: integer
      is_admin:
        type: boolean
      password_hash:
        type: string
      username:
        type: string
    type: object
...
  handler.UserAuthParams:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by comparing the submitted password to the
        hashed password stored in the database
      parameters:
      - description: Login Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserAuthParams'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token response
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid login credentials
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User login
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      description: Creates a new user in the database with a hashed password using
        the Argon2id algorithm
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/db.CreateUserParams'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully created
          schema:
            type: string
        "400":
          description: User already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes a user by ID from the database, requires admin privileges
        or account owner
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
        "400":
          description: Invalid user ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete user
      tags:
      - users
  /users/{username}:
    get:
      description: Fetches a user by their username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: User not found
          schema:
            type: string
      summary: Get user
      tags:
      - users
  /users/{username}/password:
    post:
      consumes:
      - application/json
      description: Hashes and updates the password for an existing user, requires
        admin privileges or account owner
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Password Update Request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/db.ChangePasswordParams'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            type: string
        "400":
          description: Invalid input or password criteria not met
          schema:
            type: string
        "500":
          description: Internal server error during password update
          schema:
            type: string
      summary: Update password
      tags:
      - users
...
swagger: "2.0"